fix: 필요한 기능만 남기고 나머지 삭제
