+++ src/main/java/com/example/demo/ai/controller/GptAssistantTestController.java
@@ -1,44 +1,56 @@
 package com.example.demo.ai.controller;
 
 
-
-import com.example.demo.ai.dto.assistant.GetAssistantResDto;
-import com.example.demo.ai.dto.assistant.GptApiCreateAssistantResDto;
 import com.example.demo.ai.dto.assistant.v2.Assistant;
 import com.example.demo.ai.dto.assistant.v2.AssistantCreateRequest;
-import com.example.demo.ai.dto.file.FileData;
-import com.example.demo.ai.dto.file.FileDelete;
-import com.example.demo.ai.dto.file.FileList;
-import com.example.demo.ai.dto.messages.v2.messages.Message;
-import com.example.demo.ai.dto.messages.CreateMessageResDto;
-import com.example.demo.ai.dto.messages.GetMessagesResDto;
-import com.example.demo.ai.dto.run.RunCreateRequest;
-import com.example.demo.ai.dto.run.v2.Run;
-import com.example.demo.ai.dto.thread.CreateThreadResDto;
-import com.example.demo.ai.dto.thread.DeleteThreadResDto;
+import com.example.demo.ai.dto.run.CreateThreadAndRunRequest;
+import com.example.demo.ai.dto.run.Run;
 import com.example.demo.ai.service.GptAssistantApiService;
-
-import com.example.demo.ai.service.GptAssistantApiServiceV2;
 import com.example.demo.ai.service.GptService;
-import com.example.demo.ai.service.dto.DeleteAssistantResDto;
-import jakarta.servlet.http.HttpServletRequest;
+import com.example.demo.ai.service.dto.DailyCodyResDto;
+import com.example.demo.weather.dto.fcst.FcstItem;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.util.Optional;
+import java.util.ArrayList;
+import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
 
 public class GptAssistantTestController {
-    private final GptAssistantApiService gptAssistantApiService;
-    private final GptAssistantApiServiceV2 gptAssistantApiServiceV2;
-
     private final GptService gptService;
+    private final GptAssistantApiService gptAssistantApiServiceV2;
+    @PostMapping("/v1/daily-cody")
+    public void dailyCody(
+            @RequestBody
+            CreateThreadAndRunRequest dto
+    ) {
 
+        List<FcstItem> fcstItemList = new ArrayList<>();
+        for (int i = 0; i < 10; i++) {
+            FcstItem fcstItem = new FcstItem();
+            fcstItem.setBaseDate("20220423");
+            fcstItem.setBaseTime("0500");
+            fcstItem.setCategory("none");
+            fcstItem.setFcstDate("20220423");
+            fcstItem.setFcstTime(i + ":00");
+            fcstItem.setFcstValue(i + 1 + "");
+            fcstItem.setNx(i);
+            fcstItem.setNy(i);
+            fcstItemList.add(
+                    fcstItem
+            );
+        }
 
-    @PostMapping("/v1/assistants")
+        DailyCodyResDto dailyCodyResDto = gptService.generateDailyCodyCategory(fcstItemList);
+        for (String item : dailyCodyResDto.getCategories()
+        ) {
+            System.out.println(item + "카테고리");
+        }
+    }
+
+    @PostMapping("/v2/assistants")
     public Assistant createAssistant(
             @RequestBody
             AssistantCreateRequest dto
@@ -49,116 +61,22 @@ public class GptAssistantTestController {
         return gptAssistantApiServiceV2.createAssistantAPI(dto,"fashion","0.0.1");
     }
 
-    @GetMapping("/v1/assistants/{assistantId}")
-    public Optional<GetAssistantResDto.Data> getAssistant(
-            @PathVariable("assistantId")
-            String assistantId
-    ) {
-        return gptAssistantApiService.getAssistantAPI(assistantId);
-    }
 
-    @GetMapping("/v1/assistants")
-    public GetAssistantResDto getAssistants() {
-        return gptAssistantApiService.getAssistantsAPI();
-    }
-
-    @DeleteMapping("/v1/assistants/{assistantId}")
-    public DeleteAssistantResDto deleteAssistants(@PathVariable("assistantId") String assistantId) {
-        return gptAssistantApiService.deleteAssistantAPI(assistantId);
-    }
-
-    @GetMapping("/v1/threads/{threadId}")
-    public CreateThreadResDto getThread(@PathVariable("threadId") String threadId) {
-        return gptAssistantApiService.getThread(threadId);
-    }
-
-    @PostMapping("/v1/threads")
-    public CreateThreadResDto createThread(
+    @PostMapping("/v2/threads/runs")
+    public Run createThreadAndRun(
             @RequestBody
-            RunCreateRequest dto
-    ) {
-        return gptAssistantApiService.createThreadAPI(dto.getAssistantId());
-    }
-
-    @DeleteMapping("/v1/threads/{threadId}")
-    public DeleteThreadResDto deleteThreadResDto(@PathVariable("threadId") String threadId) {
-        return gptAssistantApiService.deleteThreadAPI(threadId);
-    }
-
-    @PostMapping("/v1/threads/{threadId}/messages")
-    public Message createMessage(
-            @PathVariable("threadId")
-            String threadId,
-            @RequestBody
-            Message dto
-    ) {
-        return gptAssistantApiService.createMessageAPI(dto.getRole(), dto.getContent().stream().findFirst().get().getText().getValue(), threadId);
-    }
-
-    @GetMapping("/v1/threads/{threadId}/messages/{messageId}")
-    public CreateMessageResDto getMessage(
-            @PathVariable("threadId")
-            String threadId,
-            @PathVariable("messageId")
-            String messageId
-    ) {
-        return gptAssistantApiService.getMessageAPI(threadId, messageId);
-    }
-
-    @GetMapping("/v1/threads/{threadId}/messages")
-    public GetMessagesResDto getMessages(
-            @PathVariable("threadId")
-            String threadId
-    ) {
-        return gptAssistantApiService.getMessagesAPI(threadId);
-    }
-
-    @PostMapping("/v1/threads/{threadId}/runs")
-    public Run createRun(
-            @PathVariable("threadId")
-            String threadId,
-            @RequestBody
-            RunCreateRequest dto
-    ) {
-        return gptAssistantApiService.run(threadId, dto.getAssistantId());
-    }
-
-    @GetMapping("/v1/threads/{threadId}/runs/{runId}")
-    public Run getRun(
-            @PathVariable("threadId")
-            String threadId,
-            @PathVariable("runId")
-            String runId
-    ) {
-        return gptAssistantApiService.getRun(threadId, runId);
-    }
-
-
-
-
-    @PostMapping("/v1/files")
-    public FileData uploadFile(
-            @RequestParam("uploadFile")
-            MultipartFile uploadFile,
-            HttpServletRequest req
+            CreateThreadAndRunRequest dto
 
     ) {
-        return gptAssistantApiService.fileUploadAPI(uploadFile);
-    }
-
-    @GetMapping("/v1/files")
-    public Optional<FileList> getFiles() {
-        return gptAssistantApiService.getfiles();
-    }
-
-    @DeleteMapping("/v1/files/{fileId}")
-    public Optional<FileDelete> deleteFile(
-            @PathVariable("fileId")
-            String fileId
-    ) {
-        return gptAssistantApiService.deleteFile(fileId);
+        return gptAssistantApiServiceV2.createThreadAndRun(dto);
     }
 }
+
+// TODO: 파일 업로드
+// TODO: 백터스토어 생성
+// TODO: 업로드한 파일 백터스토어에 추가
+// TODO: 실행시 해당 백터 스토어 id 추가하여 전송 혹은 어시스턴트 생성시 백터스토어 아이디 추가하여 생성
+
 
 /**
  * thread_8EZ9Fm2oxQI0iBuEiK9SX1CG
