+++ src/main/java/com/example/demo/ai/dto/assistant/Assistant.java
