--- src/main/java/com/example/demo/ai/dto/assistant/GptApiCreateAssistantReqDto.java
@@ -1,18 +1,0 @@
-package com.example.demo.ai.dto.assistant;
-
-import com.example.demo.ai.dto.assistant.v2.Tool;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.List;
-
-@Getter
-@Setter
-@Builder
-public class GptApiCreateAssistantReqDto {
-    private String instructions;
-    private String name;
-    private String model;
-    private List<Tool> tools;
-}
