+++ src/main/java/com/example/demo/ai/dto/assistant/ToolsResources.java
