+++ src/main/java/com/example/demo/ai/dto/assistant/VectorStore.java
