--- src/main/java/com/example/demo/ai/dto/messages/CreateThreadAndRunReqDto.java
@@ -1,31 +1,0 @@
-package com.example.demo.ai.dto.messages;
-
-import com.example.demo.ai.dto.messages.v2.messages.Message;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.*;
-
-import java.util.List;
-
-@Getter
-@Setter
-@Builder
-@RequiredArgsConstructor
-@AllArgsConstructor
-public class CreateThreadAndRunReqDto {
-    @JsonProperty("assistant_id")
-    private String assistantId;
-
-    private Thread thread;
-
-
-    @Getter
-    @Setter
-    @Builder
-    @RequiredArgsConstructor
-    @AllArgsConstructor
-    public static class Thread {
-        @JsonProperty("messages")
-        private List<Message> messages;
-    }
-
-}
