+++ src/main/java/com/example/demo/ai/dto/messages/content/Annotation.java
@@ -1,6 +1,5 @@
-package com.example.demo.ai.dto.messages.v2.messages.content;
+package com.example.demo.ai.dto.messages.content;
 
-import com.example.demo.ai.dto.messages.v2.messages.content.FileCitation;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
