+++ src/main/java/com/example/demo/ai/dto/messages/content/FileCitation.java
@@ -1,4 +1,4 @@
-package com.example.demo.ai.dto.messages.v2.messages.content;
+package com.example.demo.ai.dto.messages.content;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
