+++ src/main/java/com/example/demo/ai/dto/run/CreateThreadAndRunRequestThread.java
@@ -1,15 +1,13 @@
-package com.example.demo.ai.dto.run.v2;
+package com.example.demo.ai.dto.run;
 
 
-import com.example.demo.ai.dto.assistant.v2.ToolsResources;
-import com.example.demo.ai.dto.messages.v2.messages.MessageRequest;
+import com.example.demo.ai.dto.messages.MessageRequest;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 import java.util.List;
-import java.util.Map;
 
 @Getter
 @Builder
