--- src/main/java/com/example/demo/ai/dto/run/OneStepRunReqDto.java
@@ -1,35 +1,0 @@
-package com.example.demo.ai.dto.run;
-
-import com.example.demo.ai.dto.assistant.v2.Tool;
-import com.example.demo.ai.dto.assistant.v2.ToolsResources;
-import com.example.demo.ai.dto.messages.v2.messages.Message;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-import java.util.List;
-
-@Getter
-@Builder
-@RequiredArgsConstructor
-@AllArgsConstructor
-public class OneStepRunReqDto {
-
-    @JsonProperty("assistant_id")
-    private String assistantId;
-    private Thread thread;
-    private List<Tool> tools;
-
-    @JsonProperty("tool_resources")
-    private ToolsResources toolsResources;
-
-    @Getter
-    @Builder
-    @RequiredArgsConstructor
-    @AllArgsConstructor
-    public static class Thread {
-        private List<Message> messages;
-    }
-}
