--- src/main/java/com/example/demo/ai/dto/run/RunCreateRequest.java
@@ -1,14 +1,0 @@
-package com.example.demo.ai.dto.run;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.*;
-
-@Getter
-@Setter
-@Builder
-@RequiredArgsConstructor
-@AllArgsConstructor
-public class RunCreateRequest {
-    @JsonProperty("assistant_id")
-    String assistantId;
-}
