--- src/main/java/com/example/demo/ai/dto/thread/CreateThreadReqDto.java
@@ -1,4 +1,0 @@
-package com.example.demo.ai.dto.thread;
-
-public class CreateThreadReqDto {
-}
