--- src/main/java/com/example/demo/ai/dto/thread/DeleteThreadResDto.java
@@ -1,16 +1,0 @@
-package com.example.demo.ai.dto.thread;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@RequiredArgsConstructor
-public class DeleteThreadResDto {
-    private String id;
-    private String object;
-    private boolean deleted;
-}
