+++ src/main/java/com/example/demo/ai/repo/AssistantRepo.java
@@ -1,13 +1,12 @@
 package com.example.demo.ai.repo;
 
-import com.example.demo.ai.entity.Assistant;
+import com.example.demo.ai.entity.AssistantEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import javax.swing.text.html.Option;
 import java.util.Optional;
 
-public interface AssistantRepo extends JpaRepository<Assistant, Long> {
-  Optional<Assistant> findAssistantByName(String assistantName);
-  Optional<Assistant> findAssistantByAssistantId(String assistantId);
-  Optional<Assistant> findAssistantByAssistantTypeAndVersion(String type, String version);
+public interface AssistantRepo extends JpaRepository<AssistantEntity, Long> {
+  Optional<AssistantEntity> findAssistantByName(String assistantName);
+  Optional<AssistantEntity> findAssistantByAssistantId(String assistantId);
+  Optional<AssistantEntity> findAssistantByAssistantTypeAndVersion(String type, String version);
 }
